+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## V0.Next
+
+- Mark full subtrees for replacement.
+
+  ```python
+  import sepes
+  tree = [1, 2, [3,4]]
+  tree_= sp.AtIndexer(tree)[[False,False,True]].set(10)
+  assert tree_ == [1, 2, 10]
+  ```
+
+  i.e. Inside a mask, marking a _subtree_ mask with single bool leaf, will replace the whole subtree. In this example subtree `[3, 4]` marked with `True` in the mask is an indicator for replacement.
+
 ## v0.10.0
 
 - successor of the `jax`-specific `pytreeclass`